Cleaned compilation warnings. Fixed shared ptr usage bug in SimpleChat sample.

+++ tests/DataSerializerTest.cpp
@@ -171,7 +171,7 @@ public:
 	
 	VLE8_16_Action()
 	{
-		value = randu32();
+		value = (u16)randu32();
 		if (value > VLE8_16::maxValue)
 			value = VLE8_16::maxValue;
 	}
